Add asInverted().
